Add suspicious ip and card detection

+++ src/main/java/com/example/antifraud/Antifraudsystem/Service/SuspiciousIpService.java
@@ -1,7 +1,8 @@
 package com.example.antifraud.Antifraudsystem.Service;
 
-import com.example.antifraud.Antifraudsystem.DTO.IpDTO;
 import com.example.antifraud.Antifraudsystem.DTO.StatusDTO;
+import com.example.antifraud.Antifraudsystem.Exception.BadRequestException;
+import com.example.antifraud.Antifraudsystem.Exception.EntityAlreadyExistsException;
 import com.example.antifraud.Antifraudsystem.Exception.NotFoundException;
 import com.example.antifraud.Antifraudsystem.Model.SuspiciousIP;
 import com.example.antifraud.Antifraudsystem.Repository.SuspiciousIpRepository;
@@ -19,17 +20,34 @@ public class SuspiciousIpService {
         this.suspiciousIpRepository = suspiciousIpRepository;
     }
 
-    public SuspiciousIP addSuspiciousIP(IpDTO ipDTO) {
-        suspiciousIpRepository.save(createNewSuspiciousIPWithProvidedIp(ipDTO));
-        Optional<SuspiciousIP> savedIP = suspiciousIpRepository.findSuspiciousIPByIp(ipDTO.getIp());
-        assert savedIP.isPresent();
-        return savedIP.get();
+    public SuspiciousIP addSuspiciousIP(SuspiciousIP ip) {
+        if (isInvalidIp(ip.getIp())) {
+            throw new BadRequestException();
+        }
+        if (suspiciousIpRepository.findSuspiciousIPByIp(ip.getIp()).isPresent()) {
+            throw new EntityAlreadyExistsException();
+        }
+        suspiciousIpRepository.save(ip);
+        return suspiciousIpRepository.findSuspiciousIPByIp(ip.getIp()).get();
     }
 
-    public SuspiciousIP createNewSuspiciousIPWithProvidedIp(IpDTO ipDTO) {
-        SuspiciousIP suspiciousIP = new SuspiciousIP();
-        suspiciousIP.setIp(ipDTO.getIp());
-        return suspiciousIP;
+    private boolean isInvalidIp(String ip) {
+        String[] ipSplitByDot = ip.split("\\.");
+        if (ipSplitByDot.length != 4) {
+            return true;
+        }
+        try {
+            for (String s : ipSplitByDot) {
+                int i = Integer.parseInt(s);
+                if (i < 0 || i > 255) {
+                    return true;
+                }
+            }
+        } catch (NumberFormatException nfe) {
+            System.err.println(nfe.getMessage());
+            return true;
+        }
+        return false;
     }
 
     public List<SuspiciousIP> showAllSuspiciousIps() {
@@ -37,11 +55,14 @@ public class SuspiciousIpService {
     }
 
     public StatusDTO deleteSuspiciousIp(String ip) {
+        if (isInvalidIp(ip)) {
+            throw new BadRequestException();
+        }
         Optional<SuspiciousIP> suspiciousIPOptional = suspiciousIpRepository.findSuspiciousIPByIp(ip);
         if (suspiciousIPOptional.isEmpty()) {
             throw new NotFoundException();
         }
         suspiciousIpRepository.delete(suspiciousIPOptional.get());
-        return new StatusDTO(String.format("Ip %s successfully removed!", ip));
+        return new StatusDTO(String.format("IP %s successfully removed!", ip));
     }
 }
